Add files via upload
As I was doing this project as a part of a course, I may have added some unnecessary things like the #navbar at the start of the HTML main tag, you can remove them or tailor the code for yourself. Overall it is a straightforward and responsive web-based portfolio. Enjoy using it!